update: userProfile.ts 코드 분리
