+++ src/utils/userProfile.ts
@@ -1,22 +1,25 @@
-// 이미지 업로드 onChange
-// export const uploadImage = ({
-//   imgRef,
-//   setImgFile
-// }: {
-//   imgRef: React.RefObject<HTMLInputElement>;
-//   setImgFile: React.Dispatch<React.SetStateAction<string>>;
-// }) => {
-//   const file = imgRef.current?.files?.[0];
-//   if (file) {
-//     const reader = new FileReader();
-//     reader.readAsDataURL(file);
-//     reader.onloadend = () => {
-//       if (typeof reader.result === 'string') {
-//         setImgFile(reader.result);
-//       }
-//     };
-//   }
-// };
+import { useRef, useState } from 'react';
+
+// 이미지 업로드 -> onChange
+export const useImageUpload = () => {
+  const [imgFile, setImgFile] = useState<string>('');
+  const imgRef = useRef<HTMLInputElement>(null);
+
+  const uploadImage = () => {
+    const file = imgRef.current?.files?.[0];
+    if (file) {
+      const reader = new FileReader();
+      reader.readAsDataURL(file);
+      reader.onloadend = () => {
+        if (typeof reader.result === 'string') {
+          setImgFile(reader.result);
+        }
+      };
+    }
+  };
+
+  return { imgFile, imgRef, uploadImage };
+};
 
 // 플랫폼 클릭 시
 export const onClickPlatform = ({
@@ -39,3 +42,11 @@ export const onClickGenre = ({
 }) => {
   setGenres((prev) => (prev.includes(genre) ? prev.filter((el) => el !== genre) : [...prev, genre]));
 };
+
+// 이미지 파일 -> storage에 저장
+// export const storageUpload = async({profile_image_name, file}: {profile_image_name: string, file: File}) => {
+//   await browserClient.storage.from('profile_image').upload(profile_image_name, file, {
+//     cacheControl: 'no-store',
+//     upsert: true
+//   });
+// }
